+++ core/bench/report/src/prints.rs
@@ -120,7 +120,7 @@ impl BenchmarkReport {
             .sum()
     }
 
-    pub fn total_message_batches(&self) -> u64 {
+    pub fn total_message_batches(&self) -> usize {
         let batches = self
             .individual_metrics
             .iter()
