+++ core/binary_protocol/src/cli/binary_consumer_offsets/get_consumer_offset.rs
@@ -34,7 +34,7 @@ impl GetConsumerOffsetCmd {
         consumer_id: Identifier,
         stream_id: Identifier,
         topic_id: Identifier,
-        partition_id: u32,
+        partition_id: usize,
         kind: ConsumerKind,
     ) -> Self {
         Self {
