+++ core/binary_protocol/src/utils/mapper.rs
@@ -657,7 +657,7 @@ pub fn map_topic(payload: Bytes) -> Result<TopicDetails, IggyError> {
         max_topic_size: topic.max_topic_size,
         replication_factor: topic.replication_factor,
         #[allow(clippy::cast_possible_truncation)]
-        partitions_count: partitions.len() as u32,
+        partitions_count: partitions.len(),
         partitions,
     };
     Ok(topic)
