+++ core/common/src/commands/partitions/mod.rs
@@ -19,4 +19,4 @@
 pub mod create_partitions;
 pub mod delete_partitions;
 
-const MAX_PARTITIONS_COUNT: u32 = 1000;
+const MAX_PARTITIONS_COUNT: usize = 1000;
