+++ core/sdk/src/client_wrappers/binary_topic_client.rs
@@ -51,7 +51,7 @@ impl TopicClient for ClientWrapper {
         &self,
         stream_id: &Identifier,
         name: &str,
-        partitions_count: u32,
+        partitions_count: usize,
         compression_algorithm: CompressionAlgorithm,
         replication_factor: Option<u8>,
         message_expiry: IggyExpiry,
